requirements_cpu.txt
aiofiles==24.1.0
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
av==16.1.0
bottle==0.13.4
brotli==1.2.0
cachetools==7.0.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
clr_loader==0.2.10
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.4
ctranslate2==4.7.1
distro==1.9.0
fastapi==0.128.2
faster-whisper==1.2.1
ffmpy==1.0.0
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.62.0
googleapis-common-protos==1.72.0
gradio==6.5.1
gradio_client==2.0.3
groovy==0.1.2
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.4.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
onnxruntime==1.24.1
orjson==3.11.7
packaging==26.0
pandas==3.0.0
pefile==2024.8.26
pillow==12.1.0
proto-plus==1.27.1
protobuf==5.29.6
proxy_tools==0.1.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2026.0
pyparsing==3.3.2
pypiwin32==223
pyreadline3==3.5.4
pystray==0.23.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pythonnet==3.0.5
pytz==2025.2
pywebview==6.1
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
rsa==4.9.1
ruff==0.15.0
safehttpx==0.1.7
semantic-version==2.10.0
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
sympy==1.14.0
tenacity==9.1.3
tokenizers==0.22.2
tomlkit==0.13.3
tqdm==4.67.3
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
websockets==15.0.1
